requirements.txt
dj-database-url==0.5.0
Django==3.0.3
django-filter==2.2.0
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
whitenoise==5.0.1

